pyproject.toml
[project]
name = "weather-cli"
version = "0.1.0"
description = "A CLI tool for fetching and managing weather data"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Tejasv Maheshwari", email = "javetsm70@zohomail.in" }
]
keywords = ["weather", "cli", "openweather", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
weather = "weather_cli.cli:app"

[project.urls]
Homepage = "https://github.com/Javets70/weather-cli"
Repository = "https://github.com/Javtes70/weather-cli"
Issues = "https://github.com/Javts70/weather-cli/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
